% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_at.R
\name{Create_AT_Data}
\alias{Create_AT_Data}
\title{Create_AT_Data}
\usage{
Create_AT_Data(data_at_sit, data_at_incidentes, data_dc_trocos_at, data_meteo_anual,ano_inicio_dados,ano_inicio_estimativa,ano_fim_estimativa)
}
\arguments{
\item{data_at_sit}{Base de dados extraida do SIT, que contem informacoes tecnicas das linhas AT (para ver qual a estrutura que deve ter o ficheiro, correr o exemplo)}

\item{data_at_incidentes}{Base de dados extraida do CRI, que contem informacoes sobre os incidentes que ocorreram nas linhas AT (para ver qual a estrutura que deve ter o ficheiro, correr o exemplo)}

\item{data_dc_trocos_at}{Base de dados que contem informacao sobre a localizacao dos trocos AT e respectivas variaveis associadas a essa mesma localizacao (para ver qual a estrutura que deve ter o ficheiro, correr o exemplo)}

\item{data_meteo_anual}{Base de dados que contem informacao sobre dados meteorologicos anuais por distrito}

\item{ano_inicio_dados}{Ano mais antigo relativamente ao qual temos informacao tecnica e dos incidentes que ocorreram nas linhas}

\item{ano_inicio_estimativa}{Ano mais antigo em relacao ao qual se pretende estimar as falhas}

\item{ano_fim_estimativa}{Ano mais recente em relacao o qual se pretende estimar as falhas}
}
\value{
Dataset com a informação dos segmentos e variaveis utilizadas na analise dos mesmos
}
\description{
Criacao dos segmentos AT e das respectivas variaveis utilizadas na analise dos mesmos
}
\examples{
#Ler os dados a partir de um dado ficheiro
data_at_sit <- iHEAD::data_at_sit
data_at_incidentes <- iHEAD::data_at_incidentes
data_dc_trocos_at <- iHEAD::data_dc_trocos_at
data_meteo_anual <- iHEAD::data_meteo_anual
ano_inicio_dados <- 2009
ano_inicio_estimativa <- 2012
ano_fim_estimativa <- 2018

#Chamar a funcao de calculo
resultados <- Create_AT_Data(data_at_sit, data_at_incidentes, data_dc_trocos_at, data_meteo_anual, ano_inicio_dados, ano_inicio_estimativa, ano_fim_estimativa)

}
