% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_mt.R
\name{Create_MT_Data}
\alias{Create_MT_Data}
\title{Create_MT_Data}
\usage{
Create_MT_Data(data_mt_sit, data_mt_incidentes, data_dc_trocos_mt, data_meteo_anual,ano_inicio_dados,ano_inicio_estimativa,ano_fim_estimativa)
}
\arguments{
\item{data_mt_sit}{Base de dados extraida do SIT, que contem informacoes tecnicas das linhas MT (para ver qual a estrutura que deve ter o ficheiro, correr o exemplo)}

\item{data_mt_incidentes}{Base de dados extraida do CRI, que contem informacoes sobre os incidentes que ocorreram nas linhas MT (para ver qual a estrutura que deve ter o ficheiro, correr o exemplo)}

\item{data_dc_trocos_mt}{Base de dados que contem informacao sobre a localizacao dos trocos MT e respectivas variaveis associadas a essa mesma localizacao (para ver qual a estrutura que deve ter o ficheiro, correr o exemplo)}

\item{data_meteo_anual}{Base de dados que contem informacao sobre dados meteorologicos anuais por distrito}

\item{ano_inicio_dados}{Ano mais antigo relativamente ao qual temos informacao tecnica e dos incidentes que ocorreram nas linhas}

\item{ano_inicio_estimativa}{Ano mais antigo em relacao ao qual se pretende estimar as falhas}

\item{ano_fim_estimativa}{Ano mais recente em relacao ao qual se pretende estimar as falhas}
}
\value{
Lista de 4 datasets com a informação dos segmentos e variaveis utilizadas na analise dos mesmos. O elemento da lista com a denominacao "Total" contem todos os segmentos MT, o elemento com a denomincao "Mistos" contem todos os segmentos mistos MT, o elemento com a denominacao "Aereos" contem todos os segmentos aereos MT e o elemento com a denominacao "Subterraneos" contem todos os segmentos subterraneos MT
}
\description{
Criacao dos segmentos MT e das respectivas variaveis utilizadas na analise dos mesmos
}
\examples{
#Ler os dados a partir de um dado ficheiro
data_mt_sit <- iHEAD::data_mt_sit
data_mt_incidentes <- iHEAD::data_mt_incidentes
data_dc_trocos_mt <- iHEAD::data_dc_trocos_mt
data_meteo_anual <- iHEAD::data_meteo_anual
ano_inicio_dados <- 2009
ano_inicio_estimativa <- 2012
ano_fim_estimativa <- 2018

#Chamar a funcao de calculo
resultados <- Create_MT_Data(data_mt_sit, data_mt_incidentes, data_dc_trocos_mt, data_meteo_anual, ano_inicio_dados, ano_inicio_estimativa, ano_fim_estimativa)

}
