% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{iHEAD_master_algorithm}
\alias{iHEAD_master_algorithm}
\title{Funcao para efetuar a analise no curto-prazo ou no longo-prazo}
\usage{
iHEAD_master_algorithm(flag_tipoLinha=c("AT","MT_Aereas","MT_Mistas","MT_Subterraneas"),flag_prazo=c(1,2), niter,flag_versao=c(1,2), horizonte,
anosTeste, ano_inicio_estimativa, atualizar_modelo=c("sim","nao"), afinamento=c("sim","nao"))
}
\arguments{
\item{flag_tipoLinha}{Identificacao do tipo de linha a analisar
"AT": Linhas AT
"MT_Aereas": Linhas aereas MT
"MT_Mistas": Linhas mistas MT
"MT_Subterraneas": Linhas subterraneas MT
Passar a flag entre ""}

\item{flag_prazo}{Identifica o tipo de analise que se pretende
1: Curto-prazo: seleciona as variaveis para diferentes tipos de teste para no final o modelo escolhido ser aquele com melhor performance
2: Longo-prazo: seleciona as variaveis mais explicativas}

\item{niter}{Numero de variaveis que se pretende obter na analise a longo prazo; Para curto prazo equivale ao numero de iteracoes que pretender correr o programa, sendo que o numero selecionado sera de acordo com um criterio}

\item{flag_versao}{Identifica o tipo de versao que se pretende correr, completa ou resumida
1: versao resumida: imprime lista de variaveis seleccionadas
2: versao completa: imprime lista de variaveis seleccionadas e erros das previsoes}

\item{horizonte}{Numero de anos que se pretende prever, minimo 1 e maximo 3}

\item{anosTeste}{O numero de anos a definir para o teste, minimo 1 ano de teste e maximo 3 anos de teste}

\item{ano_inicio_estimativa}{O ano em que comeca o teste, com um minimo fixado em 2016}

\item{atualizar_modelo}{Permite utilizar um modelo previamente gerado para a previsão de falhas. Caso contrário o modelo será atualizado com os dados disponiveis.}

\item{afinamento}{Permite selecionar o nivel de afinamento desejado no modelo gerado. Ao ativar o afinamento o modelo irá demorar mais tempo a atualizar os resultados.}
}
\value{
Excel com as variaveis selecionadas e os erros aos diferentes niveis geograficos
}
\description{
Esta funcao agrega todas as funcoes desenvolvidas. Ao chamar esta funcao e possivel efetuar
a analise de curto prazo ou longo prazo conforme o especificado pelo utilizador.
}
\examples{
#Utilizar a funcao para efetuar a analise a longo-prazo, numa versao completa
lista_resultados<-iHEAD_master_algorithm("MT_Aereas",2, 10, flag_versao=1, 1, 1, 2016, atualizar_modelo="sim", afinamento="sim")
#Utilizar a funcao para efetuar a analise a curto-prazo, numa versao resumida
lista_resultados<-iHEAD_master_algorithm("AT",1, 10, flag_versao=1, 1, 1, 2017, atualizar_modelo="sim", afinamento="nao")
}
