% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{prever_numero_avarias_AT}
\alias{prever_numero_avarias_AT}
\title{Prever o numero de avarias para as linhas AT}
\usage{
prever_numero_avarias_AT(...,atualizar_modelo=c("nao","sim"),afinamento=c("nao","sim"))
}
\arguments{
\item{lista_dados}{Dados a utilizar na previsao das avarias para as Linhas AT}

\item{atualizar_modelo}{Permite utilizar um modelo previamente gerado para a previsão de falhas. Caso contrário o modelo será atualizado com os dados disponiveis.}

\item{afinamento}{Permite selecionar o nivel de afinamento desejado no modelo gerado. Ao ativar o afinamento o modelo irá demorar mais tempo a atualizar os resultados.}
}
\value{
Previsões do número de avarias para cada linha AT
}
\description{
Esta funcão permite prever o número de avarias para o horizonte (anos) especificado,
tendo por base as linhas AT guardadas na base de dados
}
\examples{
#Ler os dados apartir de um dado ficheiro

prever_numero_avarias_AT(Dados_AT,atualizar_modelo="sim",afinamento="nao")

}
