% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{prever_numero_avarias_MT}
\alias{prever_numero_avarias_MT}
\title{Prever o numero de avarias para as linhas MT}
\usage{
prever_numero_avarias_MT(...,atualizar_modelo=c("nao","sim"),afinamento=c("nao","sim"),Tipo_linha_MT=c("Aereas","Subterraneas","Mistas"))
}
\arguments{
\item{lista_dados}{Dados a utilizar na previsao das avarias para as Linhas MT e classificacao desejada}

\item{atualizar_modelo}{Permite utilizar um modelo previamente gerado para a previsão de falhas. Caso contrário o modelo será atualizado com os dados disponiveis.}

\item{afinamento}{Permite selecionar o nivel de afinamento desejado no modelo gerado. Ao ativar o afinamento o modelo irá demorar mais tempo a atualizar os resultados.}

\item{Tipo_linha_MT}{Permite selecionar qual o tipo de linha MT para a qual pretendemos prever o número de avarias (i.e. Aereas, Subterraneas ou Mistas).}
}
\value{
Previsões do número de avarias para cada linha MT, tendo por base a classificação solicitada.
}
\description{
Esta funcão permite prever o número de avarias para o horizonte (anos) especificado,
tendo por base as linhas MT guardadas na base de dados
}
\examples{
#Ler os dados apartir de um dado ficheiro

prever_numero_avarias_MT(Dados_MT,atualizar_modelo="sim",afinamento="nao",Tipo_linha_MT="Subterraneas")

}
