% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{probabilidade_avaria_AT}
\alias{probabilidade_avaria_AT}
\title{Estimar probabilidade de avaria para as linhas AT}
\usage{
probabilidade_avaria_AT(...,atualizar_modelo=c("nao","sim"),afinamento=c("nao","sim"))
}
\arguments{
\item{lista_dados}{Dados a utilizar no calculo da probabilidade de avaria para as Linhas AT}

\item{atualizar_modelo}{Permite utilizar um modelo previamente gerado parao calculo da probabilidade de falha. Caso contrário o modelo será atualizado com os dados disponiveis.}

\item{afinamento}{Permite selecionar o nivel de afinamento desejado no modelo gerado. Ao ativar o afinamento o modelo irá demorar mais tempo a atualizar os resultados.}
}
\value{
Calculo da probabilidade de falha para cada linha AT
}
\description{
Esta funcão permite estimar a probabilidade de falha,
tendo por base as linhas AT guardadas na base de dados
}
\examples{
#Ler os dados apartir de um dado ficheiro

probabilidade_avaria_AT(Dados_AT,atualizar_modelo="sim",afinamento="nao")

}
