% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{probabilidade_avaria_MT}
\alias{probabilidade_avaria_MT}
\title{Estimar probabilidade de avaria para as linhas MT}
\usage{
probabilidade_avaria_MT(...,atualizar_modelo=c("nao","sim"),afinamento=c("nao","sim"),Tipo_linha_MT=c("Aereas","Subterraneas","Mistas"))
}
\arguments{
\item{lista_dados}{Dados a utilizar no calculo da probabilidade de avaria para as Linhas MT e classificacao desejada}

\item{atualizar_modelo}{Permite utilizar um modelo previamente gerado parao calculo da probabilidade de falha. Caso contrário o modelo será atualizado com os dados disponiveis.}

\item{afinamento}{Permite selecionar o nivel de afinamento desejado no modelo gerado. Ao ativar o afinamento o modelo irá demorar mais tempo a atualizar os resultados.}

\item{Tipo_linha_MT}{Permite selecionar qual o tipo de linha MT para a qual pretendemos estimar a probabilidade de avaria (i.e. Aereas, Subterraneas ou Mistas).}
}
\value{
Calculo da probabilidade de falha para cada linha MT
}
\description{
Esta funcão permite estimar a probabilidade de falha,
tendo por base as linhas MT guardadas na base de dados
}
\examples{
#Ler os dados apartir de um dado ficheiro

probabilidade_avaria_MT(Dados_MT,atualizar_modelo="sim",afinamento="nao",Tipo_linha_MT="Subterraneas")

}
