% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{processamento_dados}
\alias{processamento_dados}
\title{Processamento dos dados juntamente com feature engineering}
\usage{
pre_processamento_dados(flag_tipoLinha=c("AT","MT_Aereas","MT_Mistas","MT_Subterraneas"), horizonte=c(1,2,3), anosTeste=c(1,2,3), ano_inicio_estimativa=c(2016,sem limite))
Esta funcao so deve ser corrida depois de correr (sem chamar) a funcao xxxxx DO MIGUEL.
}
\arguments{
\item{flag_tipoLinha}{Identificacao do tipo de linha a analisar
"AT": Linhas AT
"MT_Aereas": Linhas aereas MT
"MT_Mistas": Linhas mistas MT
"MT_Subterraneas": Linhas subterraneas MT
Passar a flag entre ""}

\item{horizonte}{Numero de anos que se pretende prever, minimo 1 e maximo 3}

\item{anosTeste}{O numero de anos a definir para o teste, minimo 1 ano de teste e maximo 3 anos de teste}

\item{ano_inicio_estimativa}{O ano em que comeca o teste, com um minimo fixado em 2016}
}
\value{
Dataset relativo ao treino e ao teste com os dados processados devidamente processados
O vetor das variaveis numericas que contem as variaveis candidatas na selecao, util apenas para alimentar a funcao 'selecao_variaveis';
A flag que indica o tipo de linha, util apenas para alimentar a funcao 'selecao_variaveis'.
}
\description{
Esta funcao ira fazer o pre-processamento de dados que incluiu conversao, padronizacao, remocao e definicao de variaveis.
A funcao tambem dividie o conjunto de dados para treino e teste
}
\examples{
#Utilizar a funcao para processar os dados relativos as linhas subterraneas MT, para um horizonte de um ano, com dois anos de teste comecar em 2017.
Isto siginifica que se pretende prever as falhas em 2017 e em 2018 para linhas subterraneas MT.
dados<-pre_processamento_dados("MT_Subterraneas",1,2,2017)
}
