% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHEAD_previsao.R
\name{selecao_variaveis}
\alias{selecao_variaveis}
\title{Selecao_variaveis de variaveis significativas}
\usage{
selecao_variaveis(flag_prazo=c(1,2), niter=c(1, sem limite))
A funcao so pode ser corrido depois de se correr (nao chamar) a funcao "processar_dados"
}
\arguments{
\item{lista_dados}{Dados a utilizar no processo de selecao de variaveis}

\item{flag_tipoLinha}{Identificacao do tipo de linha a analisar
"AT": Linhas AT
"MT_Aereas": Linhas aereas MT
"MT_Mistas": Linhas mistas MT
"MT_Subterraneas": Linhas subterraneas MT
Passar a flag entre ""}

\item{flag_prazo}{Identifica o tipo de analise que se pretende
1: Curto-prazo: seleciona as variaveis para diferentes tipos de teste para no final o modelo escolhido ser aquele com melhor performance
2: Longo-prazo: seleciona as variaveis mais explicativas}

\item{niter}{Numero de variaveis que se pretende obter na analise a longo prazo; Para curto prazo equivale ao numero de iteracoes que pretender correr o programa, sendo que o numero selecionado sera de acordo com um criterio}

\item{flag_versao}{Identifica o tipo de versao que se pretende correr, completa ou resumida
1: versao resumida: imprime lista de variaveis seleccionadas
2: versao completa: imprime lista de variaveis seleccionadas e erros das previsoes}

\item{horizonte}{Numero de anos que se pretende prever, minimo 1 e maximo 3}

\item{anosTeste}{O numero de anos a definir para o teste, minimo 1 ano de teste e maximo 3 anos de teste}

\item{ano_inicio_estimativa}{O ano em que comeca o teste, com um minimo fixado em 2016}
}
\value{
Excel com as variaveis selecionadas e os erros aos diferentes niveis geograficos
}
\description{
Esta funcao ira fazer a selecao das variaveis mais significativas consoante o tipo de analise selecionada
}
\examples{
#Utilizar a funcao para selecionar as variaveis numa analise a longo-prazo, numa versao completa
selecao_variaveis(2,15,2)
#Utilizar a funcao para selecionar as variaveis numa analise a curto-prazo, numa versao resumida
selecao_variaveis(1,10,1)
}
